+++ src/main/java/com/example/StudyFriends/model/Skill.java
@@ -25,7 +25,7 @@ public class Skill {
     @Column(name = "level")
     private Integer level;
     @Column(name = "exp_amount")
-    private Integer expAmount;
+    private Double expAmount;
 
     @Column(name = "is_active")
     private Boolean isActive;
